--- src/actions/constants.js
@@ -1,2 +1,0 @@
-export const LOAD_IMAGES_SUCCESS = 'LOAD_IMAGES_SUCCESS';
-export const LOAD_IMAGE_SUCCESS = 'LOAD_IMAGE_SUCCESS';
