+++ src/reducers/imageReducers.js
@@ -1,5 +1,5 @@
 import initialState from './initialState';
-import * as types from '../actions/actionTypes';
+import * as types from '../constants/actionTypes';
 
 export default function (state = initialState.images, action) {
   switch (action.type) {
